# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Fast snippet sharing platform - API documentation",
        "title": "Gisty API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/huylvt/gisty",
            "email": "support@gisty.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/health": {
            "get": {
                "description": "Check if the service is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Service is healthy",
                        "schema": {
                            "$ref": "#/definitions/handler.HealthResponse"
                        }
                    }
                }
            }
        },
        "/pastes": {
            "post": {
                "description": "Create a new code/text snippet with optional expiration and syntax highlighting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pastes"
                ],
                "summary": "Create a new paste",
                "parameters": [
                    {
                        "description": "Paste content and options",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreatePasteRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Paste created successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.CreatePasteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request (empty content, invalid syntax_type, invalid expires_in)",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "413": {
                        "description": "Content too large (max 1MB)",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service temporarily unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pastes/{id}": {
            "get": {
                "description": "Retrieve a paste's content and metadata by its short ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pastes"
                ],
                "summary": "Get a paste by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "xK9a2B",
                        "description": "Paste short ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paste retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.GetPasteResponse"
                        }
                    },
                    "400": {
                        "description": "Missing paste ID",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Paste not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "Paste has expired",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a paste by its short ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pastes"
                ],
                "summary": "Delete a paste",
                "parameters": [
                    {
                        "type": "string",
                        "example": "xK9a2B",
                        "description": "Paste short ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Paste deleted successfully"
                    },
                    "400": {
                        "description": "Missing paste ID",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Paste not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.CreatePasteRequest": {
            "type": "object",
            "required": [
                "content"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "example": "console.log('Hello, World!')"
                },
                "expires_in": {
                    "type": "string",
                    "example": "1h"
                },
                "is_private": {
                    "type": "boolean",
                    "example": false
                },
                "syntax_type": {
                    "type": "string",
                    "example": "javascript"
                }
            }
        },
        "handler.CreatePasteResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string",
                    "example": "2024-01-15T15:00:00Z"
                },
                "short_id": {
                    "type": "string",
                    "example": "xK9a2B"
                },
                "url": {
                    "type": "string",
                    "example": "http://localhost:8080/xK9a2B"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Paste not found"
                },
                "max_size": {
                    "type": "string",
                    "example": "1MB"
                }
            }
        },
        "handler.GetPasteResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "console.log('Hello, World!')"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T14:00:00Z"
                },
                "expires_at": {
                    "type": "string",
                    "example": "2024-01-15T15:00:00Z"
                },
                "short_id": {
                    "type": "string",
                    "example": "xK9a2B"
                },
                "syntax_type": {
                    "type": "string",
                    "example": "javascript"
                }
            }
        },
        "handler.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-15T14:00:00Z"
                }
            }
        }
    }
}